Refactor: Rename topic_dynamic.py
